+++ src/app/AppLayout.tsx
@@ -39,6 +39,9 @@ export default function AppLayout() {
               <NavLink to={`/projects/${projectId}/bible`} className={navClass}>
                 Story Bible
               </NavLink>
+              <NavLink to={`/projects/${projectId}/timeline`} className={navClass}>
+                Timeline
+              </NavLink>
               <NavLink to={`/projects/${projectId}/chat`} className={navClass}>
                 Story Chat
               </NavLink>